refactor/#114: 뷰의 책임 분리를 위해 RotationCamera 구현
